refactor: delete page service
- Page service functions were moved to more appropriated existing files.
- Token validation has been moved to Router class.
- Delete session methods has been moved to user services.
- Dead import, unnecessary construct call and array return referencing 'pageService' have all been removed.
